move assimilation classes in separated folder
