+++ CSharpSourceCode/CampaignSupport/Assimilation/AssimilationCampaignBehavior.cs
@@ -4,13 +4,9 @@ using System;
 using System.Collections.Generic;
 using TaleWorlds.CampaignSystem;
 using TaleWorlds.CampaignSystem.Actions;
-using TaleWorlds.CampaignSystem.ViewModelCollection.Map;
-using TaleWorlds.Core;
 using TaleWorlds.Localization;
-using TaleWorlds.SaveSystem;
-using TOW_Core.CampaignSupport.SettlementComponents;
 
-namespace TOW_Core.CampaignSupport.CampaignBehaviors
+namespace TOW_Core.CampaignSupport.Assimilation
 {
     public class AssimilationCampaignBehavior : CampaignBehaviorBase
     {
@@ -76,81 +72,5 @@ namespace TOW_Core.CampaignSupport.CampaignBehaviors
 
 
         private List<AssimilationComponent> _assimilationComponents = new List<AssimilationComponent>();
-    }
-
-    public class SettlementCultureChangedNotificationItemVM : MapNotificationItemBaseVM
-    {
-        public SettlementCultureChangedNotificationItemVM(SettlementCultureChangedMapNotification data) : base(data)
-        {
-            this._settlement = data.ConvertedSettlement;
-            this._culture = data.NewCulture;
-            base.NotificationIdentifier = "settlementownerchanged";
-        }
-
-        public override void OnFinalize()
-        {
-            base.OnFinalize();
-            CampaignEvents.OnSettlementOwnerChangedEvent.ClearListeners(this);
-        }
-
-
-        private Settlement _settlement;
-
-        private CultureObject _culture;
-    }
-
-    public class SettlementCultureChangedLogEntry : LogEntry
-    {
-        public SettlementCultureChangedLogEntry(Settlement settlement, CultureObject culture)
-        {
-            this.ConvertedSettlement = settlement;
-            this.NewCulture = culture;
-        }
-
-        public override CampaignTime KeepInHistoryTime
-        {
-            get
-            {
-                return CampaignTime.Weeks(1f);
-            }
-        }
-
-        [SaveableField(10)]
-        public readonly Settlement ConvertedSettlement;
-
-        [SaveableField(20)]
-        public readonly CultureObject NewCulture;
-    }
-
-    public class SettlementCultureChangedMapNotification : InformationData
-    {
-        public SettlementCultureChangedMapNotification(Settlement settlement, CultureObject culture, TextObject description) : base(description)
-        {
-            this.ConvertedSettlement = settlement;
-            this.NewCulture = culture;
-        }
-
-
-        [SaveableProperty(3)]
-        public Settlement ConvertedSettlement { get; private set; }
-
-        [SaveableProperty(4)]
-        public CultureObject NewCulture { get; private set; }
-
-        public override TextObject TitleText
-        {
-            get
-            {
-                return new TextObject("{=3NCExCi1}Area Culture Converted", null);
-            }
-        }
-
-        public override string SoundEventPath
-        {
-            get
-            {
-                return "event:/ui/notification/peace";
-            }
-        }
     }
 }
