+++ CSharpSourceCode/CampaignSupport/Assimilation/SettlementCultureChangedMapNotification.cs
@@ -1,0 +1,39 @@
+﻿using TaleWorlds.CampaignSystem;
+using TaleWorlds.Core;
+using TaleWorlds.Localization;
+using TaleWorlds.SaveSystem;
+
+namespace TOW_Core.CampaignSupport.Assimilation
+{
+    public class SettlementCultureChangedMapNotification : InformationData
+    {
+        public SettlementCultureChangedMapNotification(Settlement settlement, CultureObject culture, TextObject description) : base(description)
+        {
+            this.ConvertedSettlement = settlement;
+            this.NewCulture = culture;
+        }
+
+
+        [SaveableProperty(3)]
+        public Settlement ConvertedSettlement { get; private set; }
+
+        [SaveableProperty(4)]
+        public CultureObject NewCulture { get; private set; }
+
+        public override TextObject TitleText
+        {
+            get
+            {
+                return new TextObject("{=3NCExCi1}Settlement Culture Converted", null);
+            }
+        }
+
+        public override string SoundEventPath
+        {
+            get
+            {
+                return "event:/ui/notification/peace";
+            }
+        }
+    }
+}
